Abstract dependency on Client away from public Flow Control classes by using an adapter for the callback interface
